+++ R/preview.R
@@ -1,15 +1,15 @@
 #' Open site in browser
 #'
-#' `preview_site()` opens your pkgdown site in your browser. pkgdown has been
-#' carefully designed to work even when served from the file system like
-#' this; the only part that doesn't work is search. You can use `servr::httw("docs/")`
-#'  to create a server to make search work locally.
+#' `preview_site()` opens your pkgdown site in your browser, served via a
+#' local HTTP server. This enables dynamic features such as search to work
+#' correctly in preview.
 #'
 #' @inheritParams build_article
 #' @param path Path relative to destination
 #' @export
 preview_site <- function(pkg = ".", path = ".", preview = TRUE) {
-  path <- local_path(pkg, path)
+  pkg <- as_pkgdown(pkg)
+  abs_path <- local_path(pkg, path)
 
   check_bool(preview, allow_na = TRUE)
   if (is.na(preview)) {
@@ -17,15 +17,26 @@ preview_site <- function(pkg = ".", path = ".", preview = TRUE) {
   }
 
   if (preview) {
+    root <- pkg$dst_path
+
+    if (is.null(the$server) || !identical(the$server_root, root)) {
+      the$server <- nanonext::http_server(
+        url = "http://127.0.0.1:0",
+        handlers = nanonext::handler_directory("/", root)
+      )
+      the$server$start()
+      the$server_root <- root
+    }
+
     cli::cli_inform(c(i = "Previewing site"))
-    utils::browseURL(path)
+    url <- paste0(the$server$url, "/", if (path != ".") path)
+    utils::browseURL(url)
   }
 
   invisible()
 }
 
 local_path <- function(pkg, path, call = caller_env()) {
-  pkg <- as_pkgdown(pkg)
   check_string(path, call = call)
 
   abs_path <- path_abs(path, pkg$dst_path)
@@ -33,9 +44,6 @@ local_path <- function(pkg, path, call = caller_env()) {
     cli::cli_abort("Can't find file {.path {path}}.", call = call)
   }
 
-  if (is_dir(abs_path)) {
-    abs_path <- path(abs_path, "index.html")
-  }
   abs_path
 }