+++ NEWS.md
@@ -1,5 +1,7 @@
 # pkgdown (development version)
 
+* When previewing a site, it is now served via a local http server. This enables dynamic features such as search to work correctly (@shikokuchuo, #2975).
+
 * do not autolink code that is in a link (href) in Rd files (#2972)
 
 # pkgdown 2.2.0